+++ docs/script/search_index.js
@@ -162,6 +162,12 @@ window.esdocSearchIndex = [
     "class"
   ],
   [
+    "formiojs/src/components/index.js~formiocomponentsindex",
+    "variable/index.html#static-variable-FormioComponentsIndex",
+    "<span>FormioComponentsIndex</span> <span class=\"search-result-import-path\">formiojs/src/components/index.js</span>",
+    "variable"
+  ],
+  [
     "formiojs/src/formio.form.js~formioform",
     "class/src/formio.form.js~FormioForm.html",
     "<span>FormioForm</span> <span class=\"search-result-import-path\">formiojs/src/formio.form.js</span>",
@@ -180,10 +186,17 @@ window.esdocSearchIndex = [
     "class"
   ],
   [
+<<<<<<< HEAD
     "formiojs/src/formio.pdf.builder.js~formiopdfbuilder",
     "class/src/formio.pdf.builder.js~FormioPDFBuilder.html",
     "<span>FormioPDFBuilder</span> <span class=\"search-result-import-path\">formiojs/src/formio.pdf.builder.js</span>",
     "class"
+=======
+    "formiojs/src/utils/index.js~formioutils",
+    "variable/index.html#static-variable-FormioUtils",
+    "<span>FormioUtils</span> <span class=\"search-result-import-path\">formiojs/src/utils/index.js</span>",
+    "variable"
+>>>>>>> 0c4b02527c231fedb306df632163340886a15d33
   ],
   [
     "formiojs/src/formio.wizard.js~formiowizard",
@@ -1927,13 +1940,13 @@ window.esdocSearchIndex = [
   ],
   [
     "src/components/base/base.js~basecomponent#disabled",
-    "class/src/components/base/Base.js~BaseComponent.html#instance-set-disabled",
+    "class/src/components/base/Base.js~BaseComponent.html#instance-get-disabled",
     "src/components/base/Base.js~BaseComponent#disabled",
     "member"
   ],
   [
     "src/components/base/base.js~basecomponent#disabled",
-    "class/src/components/base/Base.js~BaseComponent.html#instance-get-disabled",
+    "class/src/components/base/Base.js~BaseComponent.html#instance-set-disabled",
     "src/components/base/Base.js~BaseComponent#disabled",
     "member"
   ],
@@ -3306,6 +3319,12 @@ window.esdocSearchIndex = [
     "method"
   ],
   [
+    "src/components/datagrid/datagrid.js~datagridcomponent#checkandremoveaddbutton",
+    "class/src/components/datagrid/DataGrid.js~DataGridComponent.html#instance-method-checkAndRemoveAddButton",
+    "src/components/datagrid/DataGrid.js~DataGridComponent#checkAndRemoveAddButton",
+    "method"
+  ],
+  [
     "src/components/datagrid/datagrid.js~datagridcomponent#checkconditions",
     "class/src/components/datagrid/DataGrid.js~DataGridComponent.html#instance-method-checkConditions",
     "src/components/datagrid/DataGrid.js~DataGridComponent#checkConditions",
@@ -4072,6 +4091,12 @@ window.esdocSearchIndex = [
     "file/src/components/file/File.js.html",
     "src/components/file/File.js",
     "file"
+  ],
+  [
+    "src/components/file/file.js~filecomponent#addfileinfo",
+    "class/src/components/file/File.js~FileComponent.html#instance-method-addFileInfo",
+    "src/components/file/File.js~FileComponent#addFileInfo",
+    "method"
   ],
   [
     "src/components/file/file.js~filecomponent#addwarnings",
@@ -5325,6 +5350,12 @@ window.esdocSearchIndex = [
     "src/components/select/select.js~selectcomponent#getvalue",
     "class/src/components/select/Select.js~SelectComponent.html#instance-method-getValue",
     "src/components/select/Select.js~SelectComponent#getValue",
+    "method"
+  ],
+  [
+    "src/components/select/select.js~selectcomponent#getview",
+    "class/src/components/select/Select.js~SelectComponent.html#instance-method-getView",
+    "src/components/select/Select.js~SelectComponent#getView",
     "method"
   ],
   [
@@ -6919,13 +6950,13 @@ window.esdocSearchIndex = [
   ],
   [
     "src/formio.form.js~formioform#form",
-    "class/src/formio.form.js~FormioForm.html#instance-set-form",
+    "class/src/formio.form.js~FormioForm.html#instance-get-form",
     "src/formio.form.js~FormioForm#form",
     "member"
   ],
   [
     "src/formio.form.js~formioform#form",
-    "class/src/formio.form.js~FormioForm.html#instance-get-form",
+    "class/src/formio.form.js~FormioForm.html#instance-set-form",
     "src/formio.form.js~FormioForm#form",
     "member"
   ],
@@ -7153,13 +7184,13 @@ window.esdocSearchIndex = [
   ],
   [
     "src/formio.form.js~formioform#submission",
-    "class/src/formio.form.js~FormioForm.html#instance-set-submission",
+    "class/src/formio.form.js~FormioForm.html#instance-get-submission",
     "src/formio.form.js~FormioForm#submission",
     "member"
   ],
   [
     "src/formio.form.js~formioform#submission",
-    "class/src/formio.form.js~FormioForm.html#instance-get-submission",
+    "class/src/formio.form.js~FormioForm.html#instance-set-submission",
     "src/formio.form.js~FormioForm#submission",
     "member"
   ],